Add files via upload
HandInit demonstrates the refresh for HandJoints which updates the finger bones in real-time.
CSMachine is one of the main classes in the CSI interface I created for states and controllers.
BasicController is the base behavior for the state machine.
BaseActController represents the top-level part of the machine where you add state sections to the machine in the proc_Boot, proc_Main and etc. methods.
OperationManager is a mono behavior for handling long running co routines.
OperationAction_Download is a set of methods implementing downloads for RSS, images and other REST ops.